pyproject.toml
[project]
name = "itmo-dde"
version = "0.1.0"
description = ""
authors = [
    {name = "Arkotel"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyterlab (>=4.4.7,<5.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "wget (>=3.2,<4.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "plotly (>=6.3.1,<7.0.0)",
    "click (>=8.3.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
